Setup React and React Refresh
* add react and react-dom
* add a new index.html template with a and change the webpack config for HtmlWebpackPlugin to use it
* write a new index.js that renders our new App component into the html template in its #main div
* add some components that have render counters (using a ref so we don't trigger re-renders)
* add the react-refresh/babel plugin to the webpack config to get a better HMR experience for React
