+++ package.json
@@ -21,11 +21,14 @@
   "devDependencies": {
     "@babel/core": "^7.16.12",
     "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
     "babel-loader": "^8.2.3",
     "css-loader": "^6.4.0",
     "csv-loader": "^3.0.3",
     "html-webpack-plugin": "^5.4.0",
     "json5": "^2.2.0",
+    "react-refresh": "^0.11.0",
     "style-loader": "^3.3.0",
     "toml": "^3.0.0",
     "webpack": "^5.58.2",
@@ -35,6 +38,8 @@
     "yamljs": "^0.3.0"
   },
   "dependencies": {
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   }
 }
