+++ src/index.js
@@ -1,29 +1,6 @@
-import _ from "lodash";
-import printMe from "./print.js";
+import React from "react";
+import { render } from "react-dom";
+import { App } from "./App";
 
-function component() {
-  const element = document.createElement("div");
-  const btn = document.createElement("button");
-
-  element.innerHTML = _.join(["Hello", "webpack"], " ");
-
-  btn.innerHTML = "Click me and check the console!";
-  btn.onclick = printMe; // onclick event is bind to the original printMe function
-
-  element.appendChild(btn);
-
-  return element;
-}
-
-let element = component(); // Store the element to re-render on print.js changes
-document.body.appendChild(element);
-
-if (module.hot) {
-  module.hot.accept("./print.js", function () {
-    console.log("Accepting the updated printMe module!");
-
-    document.body.removeChild(element);
-    element = component(); // Re-render the "component" to update the click handler
-    document.body.appendChild(element);
-  });
-}
+const domContainer = document.querySelector("#main");
+render(<App />, domContainer);
