+++ webpack.config.js
@@ -1,8 +1,11 @@
 const path = require("path");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
+const ReactRefreshWebpackPlugin = require("@pmmmwh/react-refresh-webpack-plugin");
 
+const isDevelopment = process.env.NODE_ENV !== "production";
+console.log({ isDevelopment });
 module.exports = {
-  mode: "development", // prevents build error
+  mode: isDevelopment ? "development" : "production", // prevents build error
   entry: {
     index: "./src/index.js",
   },
@@ -19,17 +22,22 @@ module.exports = {
   plugins: [
     new HtmlWebpackPlugin({
       title: "WaleCMS",
+      template: "index.html",
     }),
-  ],
+    isDevelopment && new ReactRefreshWebpackPlugin(),
+  ].filter(Boolean),
   module: {
     rules: [
       {
-        test: /\.m?js$/,
+        test: /\.(js|jsx)$/,
         exclude: /(node_modules|bower_components)/,
         use: {
           loader: "babel-loader",
           options: {
-            presets: ["@babel/preset-env"],
+            presets: ["@babel/preset-env", "@babel/preset-react"],
+            plugins: [
+              isDevelopment && require.resolve("react-refresh/babel"),
+            ].filter(Boolean),
           },
         },
       },
